+++ iOSHelper/Classes/UIKit/UIImage+StackBlur.m
@@ -74,7 +74,7 @@ inline static void zeroClearInt(int* p, size_t count) { memset(p, 0, sizeof(int)
     //	return [other applyBlendFilter:filterOverlay  other:self context:nil];
 	// First get the image into your data buffer
     CGImageRef inImage = self.CGImage;
-    int nbPerCompt = CGImageGetBitsPerPixel(inImage);
+    size_t nbPerCompt = CGImageGetBitsPerPixel(inImage);
     if(nbPerCompt != 32){
         UIImage *tmpImage = [self normalize];
         inImage = tmpImage.CGImage;
@@ -95,11 +95,11 @@ inline static void zeroClearInt(int* p, size_t count) { memset(p, 0, sizeof(int)
 											 );
 	
     // Apply stack blur
-    const int imageWidth  = CGImageGetWidth(inImage);
-	const int imageHeight = CGImageGetHeight(inImage);
+    const size_t imageWidth  = CGImageGetWidth(inImage);
+	const size_t imageHeight = CGImageGetHeight(inImage);
     [self.class applyStackBlurToBuffer:m_PixelBuf
-                                 width:imageWidth
-                                height:imageHeight
+                                 width:(int)imageWidth
+                                height:(int)imageHeight
                             withRadius:inradius];
     
     // Make new image
@@ -116,7 +116,7 @@ inline static void zeroClearInt(int* p, size_t count) { memset(p, 0, sizeof(int)
 
 + (void) applyStackBlurToBuffer:(UInt8*)targetBuffer width:(const int)w height:(const int)h withRadius:(NSUInteger)inradius {
     // Constants
-	const int radius = inradius; // Transform unsigned into signed for further operations
+	const int radius = (int)inradius; // Transform unsigned into signed for further operations
 	const int wm = w - 1;
 	const int hm = h - 1;
 	const int wh = w*h;
